gravacao_teste/requirements.txt
opencv-python
numpy
Pillow
pyobjc-framework-Quartz
